Move FLCallback into api
